Indexes are automatically set an incremented when mapping in order of the Map and Reference calls.
Writing will only use the index for column ordering.

Reading will use the name if one was explicitly set. If not it will use an index if it was explicitly set. If not it will fall back to using the property name if there is a header, or the auto incremented index if there is no header or the property name didn't match a header value.