Fix for bug #423 - WriteField() & nullable types
This cahnge allows for a nullable type with a null value to be
successfully passed succesfully to WriteField<T>. Null reference error
occured because the GetType() method was being called (and is required for
unboxing) but would fail when a null field value was passed. This fix
allows typeof() to be called when the field passed in has a null value.