Add project and tests to target .NET Core
This adds a new PCL that targets the .NET Core assemblies. This will
work for .NET 4.6, ASP.NET 5, and Universal Windows Apps. It includes
the same files as the .NET 4.0 version, just compiled to target 'dotnet'
instead of '.NETFramework'.

This change also includes updates to the nuspec file so that NuGet can
correctly determine where to put files, notably all the dependencies for
libraries targeting dotnet must be explicitly listed. This is done
automatically during the build process. The 4.0-client library is also
being included in the NET 4.5 directory so it can be used correctly by
the project.json system.