adds a test showing that a custom typeconverter for Uri fails
Seems the problem has todo with the usage of an anonymous type where Uri is a constructor parameter type.
Perhaps `AutoMap` is building a map that can be used to read but we need just a IWriterConfiguration.
Finally using Uri type results in some StackOverflowException (but I hoped MyUriConverter would avoid this).