Use Encoder instead of Encoding in CsvParser
This improves the ByteCount logic when consuming surrogate characters because
the Encoder maintains state across calls. Previously the default behaviour was
that the Encoding would return the byte count having replaced the surrogate
character with U+FFFD REPLACEMENT CHARACTER because on its own a surrogate
character is invalid.

For example, when reading the UTF-16 surrogate pair \uD83D\uDE17 corresponding
to the Unicode character U+1F617, the (UTF-8) ByteCount would be 6 (the byte
count of the sequence \uFFFD\uFFFD) instead of the correct value of 4 UTF-8
bytes.