Be less hopeful about the size of the throwaway bytes buffer
An encoder with a custom EncoderReplacementFallback could easily require
a larger buffer, for example if its replacement string is "{LONG REPLACEMENT STRING}".
So the upper-bound of 16 bytes is not correct. Note that in this case
CsvParser.ByteCount would be nonsensical for input requiring the fallback, and
configuration.Encoding probably does not match the actual encoding, e.g we are reading
from a UTF-16 byte stream containing non-ASCII characters and we have set
configuration.Encoding to ASCII.

Nevertheless the safeguard is more sensible to have than not.