Ensure the ConvertToStringArgs are constructed with the correct generic type
When the parameterless Map() (returning MemberMap<object, object>) is used in
combination with Convert(ConvertToString<TClass>) then TClass is object but when
writing we invoke the delegate with a ConvertToStringArgs<T> where T is the type
of the record and not object. This fails because TClass is invariant on the
ConvertToString delegate (we cannot express it as contravariant).
So we need to construct the ConvertToStringArgs with the same generic type as
that of the ConvertToString delegate stored in WritingConvertExpression.

fixes #2118