Fix UseDefaultOnConversionFailure behavior for nullable types when default value is null
Setting a default value for a member with a nullable type with `UseDefaultOnConversionFailure` set to `true` will not use that default value when a conversion failure occurs, but throws a `TypeConverterException` instead. This is intuitive behavior and seems like a bug.

This code fixes the type comparison when the default value is set to `null` so that the code behaves like one would expect.

[discussion](https://github.com/JoshClose/CsvHelper/issues/2156#issuecomment-2245023237)
closes #2156