Remove uses of System.ValueTuple
System.ValueTuple is not included in net462. For reasons that I can't explain, I am
seeing test errors on another branch in VS (but not on the command line) relating to
missing System.Runtime.CompilerServices.Unsafe assembly. That assembly seems to be a
runtime dependency (?) of System.Memory 4.5.0 which this library references.

Bumping System.Memory to 4.5.5 appears to fix those issues, but the net462 build fails
due to missing System.ValueTuple on net462. I guess System.Memory 4.5.0 brings that in
implicitly (it is not listed as a dependency). Instead of adding an explicit reference
to the System.ValueTuple package on net462, we can just remove the few internal uses.

tl;dr this unblocks bumping the System.Memory version which fixes some local test issues